Use channels instead of single callbacks for Window events.
Motivation:
RendererBinding stores his lambda as a callback in Window.onMetricsChanged to be notified about metrics changed events. WidgetBinding has some additional logic for the same event. In dart it worked with overriding the method from RendererBinding mixin in WidgetBinding. so both logic were executed with only one callback. With our bindings delegation it will not work. WidgetBinding.handleMetricsChanged() will not be called as RendererBinding.handleMetricsChanged() was registered as a callback not WidgetsMixinWrapper's one. But we can solve it without overriding if RendererBinding and WidgetBinding both will be added as a listeners for metrics changed events. And instead of list of callbacks we can use Coroutine's channel approach, the same I used for subscribing to lifecycle events. So now Window.onMetricsChanged is a BroadcastChannel. Once metrics will be changed someone will call onMetricsChanged.offer(Unit). And Both binding's callbacks will be executed synchronously as we use launch(Unconfined) for subscribing, which means we don't need thread switches at all, just execute on the same thread offer method was called.
Also I removed Zone class, as in Dart it was like a thread as I understand. Together with callback we were storing Zone.current which means callback will be called on the same thread on which it was assigned into variable(with coroutines we will be defining CoroutineContext on the moment of subscribing(like now with Unconfined)

Test: n/a
Change-Id: I32a68742b79b9ec338ecdbc98096c566567fbae2