Make Fts3Entity and Fts4Entity require API 16
From API 11 to API 15 the SQLite version shipped with Android was 3.7.4,
while from API 16 to 19 the shipped version was 3.7.11. This means the
following FTS features are missing in APIs below 16:
1. "IF NOT EXISTS" in "CREATE VIRTUAL TABLE"
2. matchinfo option
3. languageid option
4. prefix option
5. order option

Therefore in other to nicely support FTS we've decided to require API 16
when using the FTS annotations since Room's minSDK is 14.

Updated FTS related tests to suppress them in lower OS devices.

Test: ./gradlew :room:integration-tests:testapp:cC (in both devices
      below API 16 and above it)

Change-Id: I776410fe2aef846b37354ab5b558f3184f272e92