Use provided widget params and screen size for RawImage drawing.
In the previous CL in RenderImage I used drawImage method which ignores all the params we provide for a RawImage widget. Here I ported this rich parametrizated draw method, fixed nullability of params for a widget, added ColorFilter support for Paint and added the logic for providing window metrics from SimpleFlutterView.
Now hello world image draws with respect of a screen size and provided params like fit, scale, color and alignment. It remeasures and redraws after the screen rotation. Only nine-patch drawing is not supported.

Test: checked how the image is displayed with different combination of params in the hello world app.
Change-Id: I9295fd7ba73c8b0cfed73689accef35eae1e9f38