Fixed subtle issue with animated scrolling.
The issue is that animated scrolling will unnecessarily continue 1 step
forward despite having the knowledge that it shouldn't.  For example,
if the RV is smoothscrolling in direction A, and while it is doing so
is instructed to smooth scroll in direction B, it will step one position
more in direction A before switching to direction B.

This is unneccesarily innacurate and leads to bugs when other code
is paying close attention to things like onScrollListener callbacks.

Fixed by stopping the ViewFlinger just before we are going to "start"
it via smooth scrolling.

Fixes: 113853501
Test: androidx.recyclerview.widget.RecyclerViewLayoutTest
      #smoothScrollToPosition_oppositeDirOfPrevScroll_immediatelyChangesDir

Change-Id: I7d80d719a79a423cba0142aceb9e39d066aa1a84