ExifInterface: Fix checking possible overflow
Currently, in order to prevent overflow when retrieving
thumbnail image, ExifInterface checks if the given
thumbnail offset + thumbnail length value is greater than
the length of the file. Instead of using available(), which
may require rewinding the file to the beginning, this
CL creates a getter method for the length of the file.

Bug: 71736719, Bug: 79919054
Test: Run cts ExifInterfaceTest.java
Change-Id: I1dc5824b512024f6f01d3c885d97fa42e9cf18b7