VV2: Seek to current position when view type is updated
Currently, when view type is updated while media is paused results
in showing the default image of the video view (for SurfaceView a
black screen, and for TextureView the last rendered image). To
prevent this, this CL seeks the media to the current position if the
media is paused in order to show the current position image.

Bug: 113360020
Test: manually & ./gradlew :media-widget:connectedCheck
Change-Id: I1ebb08463cff28bee8537a9c9e481564e62ced09