Create WorkerFactory and DefaultWorkerFactory.
WorkerFactory is a new interface that allows developers to
initialize Workers on their own.  This will give them more
flexibility, and the ability to use dependency injection if
they choose.  WorkManager can be initialized with a custom
WorkerFactory.

By default, we provide DefaultWorkerFactory that encapsulates
the logic previously contained in WorkerWrapper.

TODO in a future CL: Support multiple WorkerFactory instances,
                     though this is fairly easy to do at the
                     app level.
TODO in a future CL: Enforce workers are only used once.

Test: Moved tests from WorkerWrapperTest to
      DefaultWorkerFactoryTest and ran them.

Change-Id: Ia6a48355ab3fae39982b74bd67b938f2d88a3e40