Fix encoding detection crash for android.jar
Jetifier used to crash when rewriting android.jar file due to issues
with XML files headers. This is now history.

The default mechanism that we are using for encoding detection crashed
for (probably) incorrect files headers. The solution is to catch such
issue and fallback to UTF-8. I unfortunately cannot find a better
solution. Excluding android.jar file just based on the name might be
flaky and this can probably happen for other archives as well.

Bug: b/111814958
Bug: b/116408649
Test: We are testing over subset of android.jar without crashes
Change-Id: I7a407cf8487169a6c6660b5cad66f15ce11d07d6