docs: Removing unnecessary </p> tag (per style guide)
I'd added a closing paragraph tag (in aosp/785021) because I'd
mistakenly thought it was necessary to fix a doc bug; in fact, that doc
bug had already been fixed, so my tag made no difference.

Per our style guide, we do not use </p> tags in javadocs, so I'm
removing it.

Generated doc and staged to:

http://go/dac-stage/reference/androidx/webkit/ServiceWorkerClientCompat.html#shouldinterceptrequest

...to verify.

Bug: 117558703
Test: ./gradlew disttipOfTreeDocs -PofflineDocs=false
Change-Id: I4fda24628b330010e8651a646e9c4c69ffd8942f