RenderObjects as ComponentNodes representation logic
Implemented ability to display any Widget as a set of ComponentNode primitives.
To try it you can just use ContainingView(context, widget) instead of CraneView(context, widget).
Widgets works in both ways - whenever they used as a components of as before.

Every RenderObject now contains fields layoutNode and drawNode which is the components representation of it. Layouts have no drawNode as they are not performing drawing.

When we use ContainingView we disable default RenderBindings layout and draw logic and instead are calling appropriate node logic when we need to measure or draw.

Stateful widgets are also supported as after rebuilding of the widgets hierarchy the nodes hierarchy will be also recreated accordingly.

Test: tested manually with different types of widgets in the new ContainingViewActivity
Change-Id: I21cf091ac0d555313a2606bbbc281cd6363b7c51