Handle start/end positions
ExoPlayer supporst clipping via ClippingMediaSource, but its playback/seek
positions take into account clipping so we need to compensate for the
start offset.

Also add a layer of indirection to the MediaSource so that it's possible
to access the duration of the child source's timeline.

Add an assertion to cover getting the duration of clipping media items.

Bug: 80232248
Test: Ran tests in Android Studio
Change-Id: I88a4b208ec893ed7c925ca8d7d9b4da111520aad