Send idle event when fling ends at snapped position
When using a RecyclerView with a PagerSnapHelper, the idle event was not
fired if a fling ends exactly at a snapped position (i.e., the
RecyclerView needs to be scrolled 0 pixels to snap at the current
position).

The test for this bug can not use Espresso's swipe gestures, as those
gestures do not give you control over the location where the fling ends.
A custom swipe injector is implemented in SwipeToLocation.

Bug: 118663993
Test: Added test to PagerSnapHelperTest
Change-Id: I9799c807917631a4dd7bd4a60a60ab085ae50831