MediaMetadata2: Make possible sending many bitmaps through binder
This CL makes the bitmaps in MediaMetadata2 be sent through one by one
using ParcelImplListSlice, so that the metadata can be sent through
binder even though the total bitmap size exceeds the binder limit.

Also, in case of large bitmaps, it will be scaled down to pass the
binder limit.

Bug: 119293699
Test: ./media/version-compat-tests/runtest.sh 1
Change-Id: Ie19c93e63176b88f930cea6faa1c5ed22430dab1