Fix a crash bug in FragmentScenario#recreate.
FragmentScenario#recreate causes a crash if a testing Fragment doesn't have default constructor despite you provide proper FragmentFactory. This happens because Activity doesn't retain FragmentFactory after its recreation. We need to set FragmentFactory to Activity again after it is recreated right before Activity#create is called.

This CL creates FragmentFactoryHolderViewModel and keep the factory in the viewmodel. Re-set the factory to the recreated Activity in EmptyFragmentActivity.

Test: FragmentScenarioTest
Change-Id: I22a6a173875d6a47aa6d412b1f12defe1dbd6b9d