MediaRouter: Prevent unncessary onRouteChanged calls
In the previous implementation, we used getControlFilters().equals()
to check equality of the control filters of a route.
As a result, onRouteChanged is called for each route in a provider
whenever the provider publishes routes.

This CL implements a new comparison method to compare the content of control filters
not to invoke onRouteChanged callbacks unintentionally.

Bug: 119797761
Test: run support v7 demos manually
Change-Id: I70ebb8cd08c88edd630b23e2b64ab6fd16403e8a