Port of Material and related classes
From Flutter documentation about Material:
A piece of material.
The Material widget is responsible for:
1. Clipping: Material clips its widget sub-tree to the shape specified by [shape], [type], and [borderRadius].
2. Elevation: Material elevates its widget sub-tree on the Z axis by
 [elevation] pixels, and draws the appropriate shadow.
3. Ink effects: Material shows ink effects implemented by [InkFeature]s
 like [InkSplash] and [InkHighlight] below its children.

Ripples(or Inks) are drawn on top of the Material widget, so we need to port it first to have ripples implemented. Probably not all the widgets ported here will be needed in Crane world, but it is easier to start with it and then adapt them when we will be ready to do so. Also we will not use exactly Flutter Animation framework for ripples, but we can't just wait for it to be ready, so we can reuse Flutter's one as then we just need to port 1 to 1 and then will migrate to the new Animations framework when ready.

Bug: 114021677
Test: I have also all the Ripples ported locally (can't add it here as it requires Shep's GestureDetector changes) and tested them locally in CraneView. They works!
Change-Id: I8b1be1bf388923758f503ff4f054a320b8b55f04