Add FontFamily test font files
Each font contains [a,b,c,d,e,f,g,h] characters and [wide,narrow] glyph types: Each font file includes one character that is represented with
wide glyph, and others with narrow. This is used in the tests in
order to differentiate the font that is loaded for a specific 
weight/style in the FontFamily.
- 200 italic has [a] with wide glyph
- 200 regular has [b] with wide glyph
- ...
- 900 italic has [g] with wide glyph
- 900 regular has [h] with wide glyph

Test: None

Bug: 113815795
Change-Id: I621a61ae1f6e0874a1c476d4d719c1820d54dc1b