Fix MeasureBox bust parameter issue (component's overmemoize)
I investigated and it turned out it is not really overmemoize bug, it is just how R4a works.
Imagine the tree:
<CraneWrapper>
  <MeasureBox> constraints1 ->
    // first measuring block
    collect {
        <MeasureBox> constraints2 ->
           // second measuring block
    }
Once the View backed inside CraneWrapper was resized by the Activity we will get the new Constraints and we want the tree to be remeasured.
But by the current layouts design constraints are not a part of composition neither is the measuring block lambdas.
MeasureBox is a part of it, but the measuring lambda will be just stored into a field to be called later when we will have onMeasure call on MeasureBox's View.
So technically every time we have new constraints we don't need to recompose, we just need to reexecute stored measuring lambdas for our stored tree of MeasureBoxes.
And R4a will recompose only when something else was changed.

Why it wasn't working without bust:
When our root CraneWrapper(or any MeasureBox) was resized by the system we were recomposing exactly this component's children.
In our example is the first MeasureBox(it is a child of CraneWrapper).
Then R4a applies optimizations and understands the inner(second) MeasureBox didn't change. So R4a will not call compose() on it.
But our first MeasureBox created a list of measureBoxes to collect children and provided it as an ambient.
If out second MeasureBox will be recomposed it will be added into this list inside it's compose method, but compose wasn't called because of optimization.
Our first MeasureBox thinks it has no children(second MeasureBox didn't add itself) so no content displayed on the screen.
Once we add "bust: Double" parameter R4a will not apply optimization as this parameter will be changed on every call. So the second MeasureBox will be added into a list of children.

We can't rely on compose to be called on every frame for every MeasureBox so we need to collect children once they added to the composition.
We can currently emulate didCommit effect with just doing it inside the first call of compose(). Unfortunately we can't currently detect when the Component was removed from composition (should be implemented in R4a as onDispose effect). So until introducing effects this solution will work incorrectly when users of MeasureBox will recompose with removing some already displayed widgets or reorder them. But I think it is still better for most of us than current unpredictable behaviour when you need to add bust param everywhere trying to make it work.

Looks like it also helped to get rid of the requirement of providing a Draw lambda as a parameter instead of ktx children. I checked and it works now.

Test: tested our current R4a test activities
Change-Id: I9398685da4f1db226216b08dc597bdd68742b41c