Fix ChangeImageTransform wrong matrix applied on API < 21 bug
There was a logic to store the current scaleType and the matrix before the animation and restore it after the animation for pre-lollipop. It works fine until we start an another concurrent transition while the previous one is still in progress. Example steps:
1) The user scheduled the new transition and changed the scaleType to the value they wants
2) The old animation canceled and the end listener is called.
3) In this listener we apply our stored scaleType and overrides the value the user provided.
4) We start a new animation and now store the wrong value - our just restored, but not the final value provided by the user
5) After the animation we restore a wrong scaleType.
To prevent this for happening we should not save and restore but reimplement the ImageView.animateTransform() method logic for API's below 21. We have to use reflection here (only before 21) to change the value of ImageView.mDrawMatrix to apply a transient matrix for a view, it will only be applied for the next drawing and not clear the user provided matrix.

Bug: 123226255
Test: added a test for this case, fixed the test to correctly obtain a drawMatrix before 4.3
Change-Id: Id9870adc59ba815bfb37cedc42426cb6921d11df