Simplify max page logic in PagedListView.
Remove notions of DEFAULT_MAX_CLICKS, which was a legacy decision when
PagedListView used to try to limit the amount of clicks on it. Now, this
functionality should be up to the application.

Instead, simply allow a maximum number of pages to be set on the
PagedListView. By default, the maximum number is UNLIMITED.

Change-Id: If3e009a5e3375e9fa05ef0a8a565af75c0e883cc
Fixes: 123350353
Test: ./gradlew car:connectedCheck