Return a CodeBlock from CodeGenScope.generate() instead of a String
Returning the CodeBlock allows for other JavaPoet methods that add code
to correctly parse and format tokens. Otherwise retuning a String
can cause a oddly code blocks that contain dollar sings to be
interpreted as a new token placeholder.

An example of this is seen when combining the internal visibility
modifier along with a suspend function. It causes the name of the method
to be append with a dollar sign and some additional name which in turns
makes room create a sqlite statement variable containing the sign. When
the odd name is used to generate the dao method implementation JavaPoet
complains about the unknown token.

Bug: 123466702
Test: ./gradlew :room:integration-tests:room-testapp-kotlin:cC
Change-Id: I4eaf999960a7db6c6fe9b0e88d32b3b2f8d93e52