Do not call protected method from constructor
This currently results in unexpected behavior. If a user overrides the
calculateSpeedPerPixel(DisplayMetrics) method, it will be invoked while
the class is a semi-initialized state. i.e., if the user's overridden
method references any parameters set in their constructor, these
parameters will not yet have been initialized. The only way a user of
this class could possibly know this is by reading the source code.

Example code that would trigger this error:

  class MyScroller extends LinearSmoothScroller {
    private final float millisPerInch;

    MyScroller(Context context) {
      super(context);
      millisPerInch =
        context.getResources()
           .getInteger(R.integer.scroll_speed_millis_per_inch);
    }

    @Override
    protected float calculateSpeedPerPixel(
        DisplayMetrics displayMetrics) {
      // At this point millisPerInch is 0!
      return millisPerInch / displayMetrics.densityDpi;
    }
  }

Test: TODO - want first to find out if this looks acceptable first.
Change-Id: Id07cc6d954cb45e2d76fd58deebcdbb995016b45