Add themes and styling for TabLayout.
A developer who takes the default Theme.Car will automatically inherit
the styling for the TabLayout but will need to use the custom view
car_tab_view.xml for each individual tab layout.

If they do not use it, the icon size will not match UX specs, but
everything else will be usable.

Bug: 123646457
Test: ./gradlew car:connectedCheck
Test: Verify on newly created demo activity.
Change-Id: I3dc4cd0eb67553010b0133416b96ef9deaf15149