Fix default tint mode for ImageView on API 21
ImageView keeps a copy of tint mode, which defaults to null, while
drawables default their tint mode to SRC_IN. On API 21 and later,
tinting an untinted image via ImageView#setImageTintList() results in
the image being tinted using SRC_IN, even though #getImageTintMode()
will return null until another mode is specified.

ImageViewCompat includes a workaround for ImageView's drawable not
updating its state after receiving a new tint list or mode, preventing
the image from being tinted until its state was updated. This workaround
was previously gated on the presence of a tint mode on the view, which
may be null even if the underlying drawable has the default SRC_IN tint
mode. This CL removes the gate on tint mode, but keeps it on the
presence of a tint list, since that is kept in sync between the drawable
and the ImageView.

Removing the gate may result in additional drawable invalidations in API
21, but provides the correct behavior. An alternative would be to
reflect on the private field ImageView#mHasDrawableTintMode, although
this would likely be slower than additional invalidations due to the
reflection overhead.

Fixes: 117428370
Test: AppCompatImageViewTest on API 21
Change-Id: I7f8619a36efd5ae3e81126e16748c1ced3bd927a