Made Checkbox toggleable and added onClick Test
Added ability to issue onClick events during tests.

This also adds support for Checkbox on press as it is needed for
testing though at this point due to additional bugs it doesn't
corectly render on screen. Tests for these cases have been added
in this CL and annotated with @Ignore for now.

Bug: 123700326
Test: androidx.ui.material.test.CheckboxUiTest
Change-Id: If3804aba3c2671bf5e49994868c5498bd4fa5e9f