Revert changes in default.config and remove task.
It turned out that we have released (as part of beta03) mappings that
were not supposed to be used externally. This is because we were actively
maintaining "dejetifyArchive" task without realizing that next jetifier
update will contain such changes. As we are done with 28.0.0 I think it
is a good time to sunset the "dejetifyArchive" target. I also reverted
all the config changes that are not supposed to go outside. Please note
that there is still "partiallyDejetifyArchive" for migration purpouses
and is still part of continuous build.

Also added ignore rule for ResultReceiver.

Bug: b/123651524
Bug: b/124291947
Test: N/A
Change-Id: I59983d25ae7e0939f2e9e54005460ad0f73beb3c