Fix FragmentNavigator singleTop operations
As singleTop operations don't end up adding
anything to the back stack, the generated
back stack name should only use the current
back stack's size (unlike an add operation,
which uses size + 1).

As part of this, ensured that we use the
popBackStack() that takes a back stack name to
avoid popping child Fragments off the back stack.

Rewrote the testSingleTop() to use Truth asserts
throughout instead of some of each.

Test: updated test
Fixes: 124294805
Change-Id: I87a41754d8718aced9a636e9b8d00fa2038137aa