Separate background to fix bug setting visibility of center_view in MCV
There are two points that control mCenterView's visibility. One is fade
in/out animation and the other is onLayout. At the end of the fade out
animation, it sets the visibility as INVISIBLE while onLayout sets the
visibility as VISIBLE since mCenterView is expected to be shown unless
mSizeType is SIZE_TYPE_FULL. onLayout may be called as the time goes so
mCurrentTime.setText() is invoked.

Bug: 127870423
Test: Run support-media-demos manually
Change-Id: I7a728e2ed5fd1052b56190b480fcd3357ce1edd8