Added capture processing to ImageCaptureUseCase
This integrates the CaptureProcessor and CaptureBundle into the
ImageCaptureUseCase so that a set of captures will be requested when a
single takePicture() is called. Then the CaptureProcessor will process
the set of resulting Images to produce a single image as output.

Now the ImageCaptureUseCase will instanitate a ProcessingImageReader if
processing is to be done or a MetadataImageReader if not. Both of them
will attach the ImageInfo to the ImageProxy.

Whenver the processing is done the ImageFormat requested will be
YUV_420_888 instead of the format chosen by the
ImageReaderFormatRecommender.

Bug: b/124531472
Test: ./gradlew camera:camera-core:build

Change-Id: I5b6d365e69115fe2db399622ba60eff111102010