Allow <activity> destinations to set an explicit package
When navigating via an action, it is sometimes useful
to be able to limit the components to a specific
package such as when using dynamic features or are
handling web URL links within your own app.

By adding an app:targetPackage attribute, we can
ensure that developers can avoid disambiguation
dialogs in those cases.

For inflating from XML, a special package name
of ${applicationId} is supported which is populated
with your application's package name.

Test: updated tests
Fixes: 110975456
Change-Id: I0f0f95efa1cc7ca5c2f26cf9bc60819cc4478a92