Let ViewPager2 be treated as a Collection by accessibility.
Semantically this makes sense since a ViewPager contains a collection of
items. Since its items already have CollectionItemInfos belonging to RV
we can extend this and attach the CollectionInfo to ViewPager2. This
will also help identify when an a11y user navigated into a pager

CollectionInfoCompat values:
rowCount and colCount correspond to the page numbers

hierarchical was introduced for nested collections in webviews and
devs shouldn't really be messing with this

selectionMode doesn't apply here and doesn't look like well-defined api.
It was introduced for TalkBack but the app doesn't use it.

Bug: 119425427
Test: Added a test to verify collectionInfo is being initialized
correctly. Ran and passed test suite.
Change-Id: I88e66121d1659fca9815304a917cfc66ae13bc3d