Add support for contextual Notification Actions.
Contextual Notification Actions are Actions that depend on the
notification content - more specifically the message of a message
notification.
Contextual Actions are only supported on Q and above. They are displayed
differently than normal actions.

Bug: 121193571
Test: create a test-app and add an action on which we call
NotificationCompat.Action.Builder.setContextual(true), and ensure that
action shows up as a Smart Action.
Change-Id: I68580e40678a553e7319aa8fd7ed2805b00ae8c9