Add initial version of findByText
As the merge semantics aren't ready yet, this CL only adds an initial
version of findByText which in reality finds the text widget instead
of the actual widget that holds the text.

In this initial form this can be used to find and tap a button
as the doClick API uses coordinates to tap on the screen.

Bug: 129678658
Test: androidx.ui.material.ButtonUiTest

Change-Id: I393398ab472a70de6bada0026f565caf14bc386a